requirements.txt
Django==1.9.7
djangorestframework==3.3.0
django-filter==0.10.0

